avoid setuptools update for macOS

+++ ci/azure/install.sh
@@ -2,15 +2,18 @@
 set -ev
 
 if [ "$INSTALL_TYPE" == "conda" ]; then
-  # Workaround: https://github.com/conda/conda/issues/9337
-  pip uninstall -y setuptools
-  conda config --set always_yes yes --set changeps1 no
-  conda install -yq setuptools
-  conda update -yq conda
-  conda install conda-build anaconda-client
-  conda config --add channels conda-forge
-  conda create -n venv --yes python=$PYTHON_VERSION pip
-  conda install -yq --name venv $DEPENDS $EXTRA_DEPENDS pytest
+
+    conda config --set always_yes yes --set changeps1 no
+    if [ "$AGENT_OS" == "Linux" ]; then
+        # Workaround: https://github.com/conda/conda/issues/9337
+        pip uninstall -y setuptools
+        conda install -yq setuptools
+    fi
+    conda update -yq conda
+    conda install conda-build anaconda-client
+    conda config --add channels conda-forge
+    conda create -n venv --yes python=$PYTHON_VERSION pip
+    conda install -yq --name venv $DEPENDS $EXTRA_DEPENDS pytest
 else
     PIPI="pip install --timeout=60 --find-links=$EXTRA_WHEELS"
 
@@ -24,8 +27,10 @@ else
     # just to check python version
     python --version
 
-    # Needed for Python 3.5 wheel fetching
-    $PIPI --upgrade pip setuptools
+    if [ "$AGENT_OS" == "Linux" ]; then
+        # Needed for Python 3.5 wheel fetching
+        $PIPI --upgrade pip setuptools
+    fi
 
     $PIPI pytest
     $PIPI numpy
